requirements.txt
aiohttp==3.6.2
python-decouple==3.3